Move the responsibility for name resolution from hdd_reader back to hdd_monitor. Hdd_monitor itself needs resolved names.
